Refactor Falstad into parsers
Eliminate some annoying code dupes in load, hopefully with all the
integration working as-is.
