+++ core/src/main/kotlin/org/eln2/parsers/falstad/components/generic/Ignore.kt
@@ -1,0 +1,8 @@
+package org.eln2.parsers.falstad.components.generic
+
+import org.eln2.parsers.falstad.CCData
+import org.eln2.parsers.falstad.IComponentConstructor
+
+class Ignore: IComponentConstructor {
+	override fun construct(ccd: CCData) {}
+}
